clean config (#39)
* remove the shared stuff from one config
* remove shared stuff from every config
