+++ models/llama-2-70b-chat/config.py
@@ -1,34 +1,17 @@
-from collections import OrderedDict
-import logging
-import re
-import time
-from transformers import LlamaTokenizer, AutoConfig, LlamaForCausalLM
-import torch
-import subprocess
-from subprocess import DEVNULL, STDOUT
-from tensorizer import TensorDeserializer
-from tensorizer.utils import no_init_or_tensor
-import os
 from dotenv import load_dotenv
-
 from src.utils import get_env_var_or_default
-
-from subclass import YieldingLlama
-
-# add parent directory to path
-import sys
 
 load_dotenv()
 
-MODEL_NAME = 'llama-2-70b-chat'
+MODEL_NAME = "llama-2-70b-chat"
 # INFERENCE CONFIGURATION
 #######################################################################
 # --------------------Notes--------------------------------------------
-# We sometimes implement inference differently for models that have not 
-# been trained/fine-tuned vs. those that have been trained/fine-tuned. We refer to the 
+# We sometimes implement inference differently for models that have not
+# been trained/fine-tuned vs. those that have been trained/fine-tuned. We refer to the
 # former as "default" and the latter as "trained". Below, you can
 # set your "default inference configuration" and your "trained
-# inference configuration". 
+# inference configuration".
 #
 # GENERAL INFERENCE CONFIGURATION
 # -------------------------------
@@ -50,12 +33,13 @@ USE_FUSED_ATTN = False
 # -------------------------------
 
 
-LOCAL_DEFAULT_INFERENCE_WEIGHTS_PATH = f"models/{MODEL_NAME}/model_artifacts/default_inference_weights"
+LOCAL_DEFAULT_INFERENCE_WEIGHTS_PATH = (
+    f"models/{MODEL_NAME}/model_artifacts/default_inference_weights"
+)
 
 REMOTE_DEFAULT_INFERENCE_WEIGHTS_PATH = get_env_var_or_default(
-    "REMOTE_DEFAULT_INFERENCE_WEIGHTS_PATH", 
+    "REMOTE_DEFAULT_INFERENCE_WEIGHTS_PATH",
     "remote/path/to/your/weights/here",
-
 )
 
 # N_SHARDS = 2
@@ -84,15 +68,17 @@ REMOTE_DEFAULT_INFERENCE_FILES_TO_DOWNLOAD += [
 LOCAL_TRAINING_WEIGHTS_PATH = f"models/{MODEL_NAME}/model_artifacts/training_weights"
 
 REMOTE_TRAINING_WEIGHTS_PATH = get_env_var_or_default(
-    var_name="REMOTE_TRAINING_WEIGHTS_PATH", 
-    default_value="remote/path/to/your/weights/here"
+    var_name="REMOTE_TRAINING_WEIGHTS_PATH",
+    default_value="remote/path/to/your/weights/here",
 )
 
-LOCAL_TRAINING_WEIGHTS_CONFIG_PATH = f"models/{MODEL_NAME}/model_artifacts/training_weights/config.json"
+LOCAL_TRAINING_WEIGHTS_CONFIG_PATH = (
+    f"models/{MODEL_NAME}/model_artifacts/training_weights/config.json"
+)
 
 REMOTE_TRAINING_WEIGHTS_CONFIG_PATH = get_env_var_or_default(
-    var_name="REMOTE_TRAINING_WEIGHTS_CONFIG_PATH", 
-    default_value="remote/path/to/your/weights/here"
+    var_name="REMOTE_TRAINING_WEIGHTS_CONFIG_PATH",
+    default_value="remote/path/to/your/weights/here",
 )
 
 N_SHARDS = 15
@@ -110,74 +96,3 @@ REMOTE_TRAINING_FILES_TO_DOWNLOAD += [
     "tokenizer.json",
     "tokenizer.model",
 ]
-
-
-# -------------------------------
-
-DEFAULT_PAD_TOKEN = "[PAD]"
-DEFAULT_EOS_TOKEN = "</s>"
-DEFAULT_BOS_TOKEN = "<s>"
-DEFAULT_UNK_TOKEN = "</s>"
-
-
-def log_memory_stuff(prompt=None):
-    """One method to barf out everything we'd ever want to know about memory"""
-    
-    if prompt is not None:
-        print(prompt)
-    os.system("nvidia-smi")
-    print(torch.cuda.memory_summary())
-
-
-def load_tokenizer():
-    """Same tokenizer, agnostic from tensorized weights/etc"""
-    tok = LlamaTokenizer.from_pretrained(TOKENIZER_PATH, cache_dir="pretrained_weights", legacy=False)
-    tok.add_special_tokens(
-        {
-            "eos_token": DEFAULT_EOS_TOKEN,
-            "bos_token": DEFAULT_BOS_TOKEN,
-            "unk_token": DEFAULT_UNK_TOKEN,
-            "pad_token": DEFAULT_PAD_TOKEN,
-        }
-    )
-    return tok
- 
-
-def load_tensorizer(
-    weights, plaid_mode: bool = True, cls: LlamaForCausalLM = YieldingLlama
-):
-    st = time.time()
-    weights = str(weights)
-
-    if 'http' in weights:
-        if not (os.path.exists(LOCAL_TRAINING_WEIGHTS_PATH)):
-            download_file(weights, LOCAL_TRAINING_WEIGHTS_PATH)
-        weights = LOCAL_TRAINING_WEIGHTS_PATH
-    
-    if not os.path.exists(LOCAL_TRAINING_WEIGHTS_CONFIG_PATH):
-        download_file(REMOTE_TRAINING_WEIGHTS_CONFIG_PATH, LOCAL_TRAINING_WEIGHTS_CONFIG_PATH)
-
-    config = AutoConfig.from_pretrained(LOCAL_TRAINING_WEIGHTS_CONFIG_PATH)
-
-    logging.disable(logging.WARN)
-    model = no_init_or_tensor(
-        lambda: cls.from_pretrained(
-            None, config=config, state_dict=OrderedDict(), torch_dtype=torch.float16
-        )
-    )
-    logging.disable(logging.NOTSET)
-
-    des = TensorDeserializer(weights, plaid_mode=plaid_mode)
-    des.load_into_module(model)
-    print(f"weights loaded in {time.time() - st}")
-    
-    # We don't know what device model was tensorized in or what dtype was used.
-    # If a GPU is available, we need to ensure that the model is on the GPU and cast to fp16.
-    if next(model.parameters()).is_cuda:
-        model = model.half()
-    else:
-        if torch.cuda.is_available():
-            model.to("cuda")
-            model = model.half()
-
-    return model
